backend/django_project/requirements/requirements-production.txt
Django>=4.0
jinja2==3.0.1
psycopg2-binary
requests
python3-memcached
pymemcache
gunicorn
django-jazzmin
django-cors-headers
ipython
uritemplate
reportlab
gunicorn
celery[redis]
autobahn
Twisted[tls,http2]
redis
Pillow
firebase_admin
djangorestframework-xml
djangorestframework
channels
channels_redis
beautifulsoup4
html5lib
django_pandas
numpy
qrcode
xlsxwriter